Implement interim compatibility with 1.33 WANObjectCache keys
WANObjectCache keys formatting changed between 1.33 and 1.37, with the
effect that e.g. invalidating a global key from 1.33 will not be
effective for 1.37 wikis that have cached the same global key. As a
workaround, add temporary support for the 1.33 key format to 1.37. This
can be switched once the migration is over.